requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
faiss-cpu==1.7.4
openai==1.3.0
langchain==0.0.339
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
tiktoken==0.5.1
PyPDF2==3.0.1
numpy==1.26.2
